תזמון צירים

+++ src/main/java/com/ashcollege/entities/UserEntity.java
@@ -10,7 +10,7 @@ public class UserEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     private String firstName;
     private String lastName;
@@ -23,10 +23,11 @@ public class UserEntity {
     public UserEntity() {
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
-    public void setId(int id) {
+
+    public void setId(Long id) {
         this.id = id;
     }
